requirements.txt
Flask==2.3.3
joblib==1.3.2
numpy==1.24.3
pandas==2.0.3
psycopg2-binary==2.9.11
python-dotenv==1.0.0
scikit-learn
